Delete styles dir. Move everything into config/ or const/
